% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_dttm.R
\name{align_dttm}
\alias{align_dttm}
\title{Align Date and Time Fields}
\usage{
align_dttm(tble, date, time, key)
}
\arguments{
\item{tble}{extraction from core with date and time of death in separate rows}

\item{date}{date column}

\item{time}{time column}

\item{key}{key linking date and time (usually episode_id)}
}
\value{
a tibble
}
\description{
This function takes two fields that are perfectly disjoint with nas and combines them into 1
This is useful where you have details like date and time of death stored in different columns
}
