% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_int.R
\name{bind_na_int}
\alias{bind_na_int}
\title{Binds processed output table and NA table}
\usage{
bind_na_int(processed_table = NULL, na_table = NULL)
}
\arguments{
\item{na_table}{}
}
\description{
Binds processed output table and NA table
}
