% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_parameters.R
\name{calc_alive_length}
\alias{calc_alive_length}
\title{Calculate Episode Length From Mortality Data}
\usage{
calc_alive_length(core, reference_table = NULL)
}
\arguments{
\item{core}{core table from \code{make_core()}}
}
\description{
Date of death and time of death are stored separately in CC-HIC. This function calculates the
date and time of death of a patient, and returns a tibble by episode_id
}
