% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_parameters.R
\name{calc_epi_length}
\alias{calc_epi_length}
\title{Calculate Episode Length}
\usage{
calc_epi_length(core_table = NULL, reference_table = NULL, na.rm = TRUE)
}
\arguments{
\item{na.rm}{logical to remove missing values}

\item{core}{core table from \code{make_core()}}
}
\description{
Calculates episode length based upon the arrival dttm into the ICU and discharge dttm.
This does not take into account any mortality data (There is an
unexplained discrepancy between discharge and the mortality fields).
All times are reported in difftime units days. If na.rm = FALSE then all values are returned,
including those case that do not have a discharge time.
}
