% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_parameters.R
\name{calc_site_occupancy}
\alias{calc_site_occupancy}
\title{Calculate Estimated Site Occupancy}
\usage{
calc_site_occupancy(episode_length_tbl = NULL, impute = TRUE)
}
\arguments{
\item{episode_length_tbl}{episode length table}

\item{episodes_tbl}{episodes table}

\item{provenance_tbl}{provenance table}
}
\value{
a table of similar structure to date_skelaton, but with estimated occupancies attached
}
\description{
Calculate Estimated Site Occupancy
}
