% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connect}
\alias{connect}
\title{Establishes a connection to the postgres database}
\usage{
connect(host = "localhost", username = "ucasper", password = "superdb",
  database = "cchic")
}
\arguments{
\item{host}{host of the database: typically localhose}

\item{database}{}
}
\description{
Establishes a connection to the postgres database
}
