% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{Describes the coverage for reporting of a variable (S3 Generic)}
\usage{
coverage(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Describes the coverage for reporting of a variable (S3 Generic)
}
