% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendarHeat.R
\name{create_calendar}
\alias{create_calendar}
\title{Create A Heat Map Calender}
\usage{
create_calendar(x = NULL, start_date = "2014-01-01",
  end_date = "2018-01-01")
}
\arguments{
\item{x}{count data to be used in the calendar}

\item{start_date}{a starting date as character vector of format YYYY/MM/DD}

\item{end_date}{an end date as character vector of format YYYY/MM/DD}
}
\value{
a tibble with correct week alignments for a calendar heatmap
}
\description{
creates the basic template for a heatmap calendar in ggplop
}
\examples{
sample_data <- create_calendar(myData, "2014-01-01", "2018-01-01")
}
