% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendarHeat.R
\name{create_grid}
\alias{create_grid}
\title{Create Calendar Grid}
\usage{
create_grid(calendar_template = NULL)
}
\arguments{
\item{calendar_template}{}
}
\value{
a grid for assembling a ggheatcal
}
\description{
Create Calendar Grid
}
