% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{empty_files}
\alias{empty_files}
\title{Empty XML Files}
\usage{
empty_files(episodes, provenance)
}
\arguments{
\item{episodes}{collected episodes postgres table}

\item{provenance}{collected provenance postgres table}
}
\value{
a tibble containing the site and filename of XML with no episodes
}
\description{
Details XML that has successfully parsed, but contains no episode level data
}
\examples{
empty_files(episodes, provenance)
}
