% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_parameters.R
\name{epi_end_alive}
\alias{epi_end_alive}
\title{Collect Episode End Datetime}
\usage{
epi_end_alive(core_table = NULL)
}
\arguments{
\item{core_table}{core table from \code{make_core()}}
}
\value{
a tibble with mandatory episode defining characteristics.
}
\description{
Collects episodes end datetime for further processing
}
\examples{
epi_end_alive(core)
}
