% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_parameters.R
\name{epi_length}
\alias{epi_length}
\title{Calculate Episode Length}
\usage{
epi_length(core, useMortality = FALSE, reference_table = NULL,
  na.rm = FALSE)
}
\arguments{
\item{core}{core table from \code{make_core()}}

\item{useMortality}{logical value for updating with mortality data}
}
\description{
Calculates length of stay and returns a tibble tagged with mortality data
}
