% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{error_summary}
\alias{error_summary}
\title{Summarise Event Level Errors}
\usage{
error_summary(errors = NULL, provenance = NULL)
}
\arguments{
\item{errors}{error table}

\item{provenance}{provenance table}
}
\value{
a tibble with summary details of errors
}
\description{
returns a summary table grouped by site with a breakdown of the errors seen.
This is for overview purposes, so scope out the magnitude of problems and help
direct xml fixes
}
\examples{
error_summary(errors, provenance)
}
