% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Extract Validated NHIC Events}
\usage{
extract(core_table = NULL, valid_episodes = NULL, qref = NULL,
  input = "NIHR_HIC_ICU_0557")
}
\arguments{
\item{core_table}{core table}

\item{valid_episodes}{an integer vector of episode IDs that have passed the validation process}

\item{qref}{the quality reference table}

\item{input}{the input variable of choice}
}
\value{
A tibble 1 row per event
}
\description{
Extracts NHIC events and labels them with the correct class for future processing
}
