% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_demographics.R
\name{extract_demographics}
\alias{extract_demographics}
\title{Extract non-time series data from CC-HIC}
\usage{
extract_demographics(metadata = NULL, events = NULL)
}
\arguments{
\item{metadata}{An uncollected postgres metadata table}

\item{events}{An uncollected postgres events table}
}
\value{
A tibble of non time series data in the same form as \code{events}
}
\description{
Extract non-time series data from CC-HIC
}
\examples{
extract_demographics(metadata, events)
}
