% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_int.R
\name{find_na_int}
\alias{find_na_int}
\title{Find episodes that did not have a record for this real}
\usage{
find_na_int(processed_table = NULL, keychain = NULL,
  provenance_table = NULL)
}
\arguments{
\item{input}{}
}
\description{
Provides a tibble of the exact same form as process dbl, but with NAs where values are
not found. This is useful in understanding how many cases do not report on a certain field
}
