% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_flags.R
\name{flag_all}
\alias{flag_all}
\title{Apply warning flags to extracted data item}
\usage{
flag_all(x, qref = NULL, los_table = NULL)
}
\arguments{
\item{x}{extracted data item}

\item{qref}{quality reference table}

\item{los_table}{episode length table}
}
\value{
a tibble with flags applied
}
\description{
Applies warning flags to an extracted HIC data item. This is a wrapper function
to flag_range, flag_bounds and flag duplicate
}
