% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_flags.R
\name{flag_bounds_2d}
\alias{flag_bounds_2d}
\title{Flag boundaries of 2d data items}
\usage{
flag_bounds_2d(x = NULL, los_table = NULL)
}
\arguments{
\item{los_table}{}
}
\description{
Flag boundaries of 2d data items
}
