% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_flags.R
\name{flag_periodicity_numeric}
\alias{flag_periodicity_numeric}
\title{Flag periodicity of 2d class}
\usage{
flag_periodicity_numeric(x, los_table = NULL)
}
\arguments{
\item{x}{data item}

\item{los_table}{episode length table}
}
\description{
Flag periodicity of 2d class
}
