% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_flags.R
\name{flag_range}
\alias{flag_range}
\title{Flag Events by Reference Range (S3 Generic)}
\usage{
flag_range(x, ...)
}
\arguments{
\item{x}{an extracted event table}

\item{...}{further arguments to pass to the S3 method}
}
\value{
a tibble of the same length as x with the following features:
\item{1}{event is a duplicate}
\item{0}{event is not a duplicate}
}
\description{
S3 generic
Flags each event as being in range (0), out of range high (+1) or out of range low (-1)
}
