% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_flags.R
\name{flag_range_numeric}
\alias{flag_range_numeric}
\title{Flag Range Numeric}
\usage{
flag_range_numeric(x = NULL)
}
\arguments{
\item{x}{}
}
\description{
The generic function for numeric type range flags
}
