% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendarHeat.R
\name{ggHeatCal}
\alias{ggHeatCal}
\title{Plot Calendar Heatmap - Ggplot layers}
\usage{
ggHeatCal(x, gridLines, Title)
}
\arguments{
\item{Title}{title of the final plot}
}
\value{
a ggplot2 object
}
\description{
layer building function of plot_heatmap
}
