% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_parameters.R
\name{invalid_months}
\alias{invalid_months}
\title{Show Invalid Months}
\usage{
invalid_months(episodes, provenance, threshold = 10)
}
\arguments{
\item{episodes}{collected episodes table}

\item{provenance}{collected provenance table}

\item{threshold}{threshold number of days to use in calculation}
}
\value{
a tibble of months that are to be excluded from future analysis
}
\description{
Determines which months are invalid based on a low contribution of data
This is based upon the long term daily average for admissions. Days that
fall below 2 SD of the long term mean are tagged. If more than the
threshold value occur in a single month, the month is removed from
further analysis
}
\examples{
invalid_months(episodes, provenance)
invalid_months(episodes, provenance, theshold = 15)
}
