% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{ks_test}
\alias{ks_test}
\title{Pairwise Kolmogorov-Smirnov Distance}
\usage{
ks_test(x)
}
\arguments{
\item{x}{extracted data item}

\item{site}{}
}
\description{
This takes a distribution from a CC-HIC data item and returns
the pairwise KS distance by site.
}
\examples{
ks_test(heart_rates)
}
