% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_core}
\alias{make_core}
\title{Produces the core table structure necessary for the data quality report}
\usage{
make_core(events, episodes, provenance)
}
\arguments{
\item{events}{events table}

\item{episodes}{episodes table}

\item{provenance}{provenance table}
}
\description{
Produces the core table structure necessary for the data quality report
}
