% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_reference}
\alias{make_reference}
\title{Make reference Table}
\usage{
make_reference(episodes, provenance)
}
\arguments{
\item{episodes}{collected episode table}

\item{provenance}{collected provenance table}
}
\value{
a tibble with episode level data with site
}
\description{
makes the reference table used as the basis for many local (non-database) functions
This is to facilitate grouping by site, as this is not present in the
episode table
}
\examples{
make_reference(episodes, provenance)
}
