% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{non_parsed_files}
\alias{non_parsed_files}
\title{Show Non Parsed XML}
\usage{
non_parsed_files(provenance = NULL)
}
\arguments{
\item{provenance}{provenance table - containing details of what was parsed}
}
\value{
a character vector of filenames
}
\description{
Returns a character vector of file names for XML that would not be parsed
}
\examples{
non_pased_files(provenance)
}
