% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendarHeat.R
\name{plot_heatcal}
\alias{plot_heatcal}
\title{Plot Calendar Heatmap}
\usage{
plot_heatcal(episodes = NULL, provenance = NULL, site = NULL,
  filename = NULL)
}
\arguments{
\item{episodes}{collected episodes table}

\item{provenance}{collected provenance table}

\item{site}{character string of hospital site to plot}

\item{filename}{optional string to save output. must include extension}
}
\value{
a plot with admission details in heatmap form
}
\description{
Plot Calendar Heatmap
}
