% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hic}
\alias{plot_hic}
\title{Plot NHIC Events}
\usage{
plot_hic(x)
}
\arguments{
\item{x}{flagged table}
}
\value{
A tibble 1 row per event
}
\description{
Plots NHIC events in a predetermined way and output to the plot folder specified
}
