% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_dttm.R
\name{process_dttm}
\alias{process_dttm}
\title{Provides a raw table for extracted datetime values}
\usage{
process_dttm(core_table = NULL, input = "NIHR_HIC_ICU_0411")
}
\arguments{
\item{input}{}

\item{outliers}{integer value for the number of standard deviations used for outlier detection}
}
\value{
A tibble of time series data in summary
}
\description{
Provides a raw table for extracted datetime values
}
\examples{
\dontrun{process_dbl(core)}
}
