% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_int.R
\name{process_int}
\alias{process_int}
\title{Provides a raw table for extracted integer values}
\usage{
process_int(core_table = NULL, input = "NIHR_HIC_ICU_0108", outliers = 3)
}
\arguments{
\item{input}{}

\item{outliers}{integer value for the number of standard deviations used for outlier detection}
}
\value{
A tibble of time series data in summary
}
\description{
This function processes the core table to pull out a specific named variable of the real class
As a demo, this runs for CRP (0557) if not supplied with defaults.
Outliers are determined by number of SD. This is a gross estimation as many variables are
not normally distributed in HIC
}
\examples{
\dontrun{process_int(core)}
}
