% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_report.R
\name{pull_cases_all}
\alias{pull_cases_all}
\title{Retrieve all unique cases}
\usage{
pull_cases_all(episodes = NULL, provenance = NULL)
}
\arguments{
\item{provenance}{}
}
\description{
Pulls unique record numbers coming through CC-HIC. This requires only that an "episode" has
been correctly parsed from the xml and added to the database and so doesn't rely on, or infer that
there is any, event level data. This is the most accurate measure of the actual number of "patients"
in the system.
}
