% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_report.R
\name{report_cases_all}
\alias{report_cases_all}
\title{Reports Case Numbers by Site}
\usage{
report_cases_all(unique_cases_tbl = NULL)
}
\arguments{
\item{unique_cases_tbl}{from \code{pull_cases_all}}
}
\description{
Reports Case Numbers by Site
}
