% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_parameters.R
\name{report_cases_byday}
\alias{report_cases_byday}
\title{Reports Case Numbers by Site and Day}
\usage{
report_cases_byday(unique_cases_tbl = NULL, bysite = NULL)
}
\arguments{
\item{unique_cases_tbl}{from \code{pull_cases_all}}

\item{site}{}
}
\value{
a tibble with the number of unique episodes admitted for a given day
}
\description{
Reports Case Numbers by Site and Day
}
\examples{
report_cases_byday(unique_cases, bysite == "UCL")
}
