% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_parameters.R
\name{report_cases_unit}
\alias{report_cases_unit}
\title{Reports Case Numbers by Unit}
\usage{
report_cases_unit(events_table = NULL, reference_table = NULL)
}
\arguments{
\item{events_table}{the events table}

\item{reference_table}{the reference table}
}
\description{
Reports on ICNARC CMP unit codes to describe the numbers of reported cases by site.
This is dependent on accurate parsing of the xml schema for hic code 0002 (INCARC CMP
unit code). And as such, may dramatically under-report. To see a comprehensive list
of acual cases, see the \code{pull_unique_cases()} function.
}
