% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_int.R
\name{report_int}
\alias{report_int}
\title{Reports Double}
\usage{
report_int(core_table = NULL, los_table = NULL, keychain = NULL,
  provenance_table = NULL, input = "NIHR_HIC_ICU_0108", outliers = 3)
}
\arguments{
\item{outliers}{}
}
\description{
Reports Double
}
