% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_parameters.R
\name{resolve_date_time}
\alias{resolve_date_time}
\title{Calculate Date and time of events}
\usage{
resolve_date_time(core, date_code, time_code)
}
\arguments{
\item{core}{core table from \code{make_core()}}
}
\value{
a function to encapsulate paired date and time variables stored separately
}
\description{
Many events in CC-HIC are stored in separate date and time columns. What is more,
this information is not always stored with consistent rules. For example,
death date and time, are stored for every patient in every episode, even though
the patient can obviously only die once. This has most likely been born out of
the ease of writing an xml extract to encompass this information. The following
are some date and time pairings that denote a singular event:
\itemize{
  \item "NIHR_HIC_ICU_0042", "NIHR_HIC_ICU_0043" - Unit Death
  \item "NIHR_HIC_ICU_0038", "NIHR_HIC_ICU_0039" - Body Removal
  \item "NIHR_HIC_ICU_0044", "NIHR_HIC_ICU_0045" - Brain stem death
  \item "NIHR_HIC_ICU_0048", "NIHR_HIC_ICU_0049" - Treatment Withdrawal
  \item "NIHR_HIC_ICU_0050", "NIHR_HIC_ICU_0051" - Discharge ready
}
This closure returns a function for the particular date time pairing that you
are interested in so that it can be incorporated into a mapping function
}
\details{
Since we are only interested in datetimes, any occurances where a date or a time
occur in isolation are dropped - This is largely because one piece of information
wihtout the other is not useful to us.
}
\examples{
resolve_death <- resolve_date_time(core, "NIHR_HIC_ICU_0042", "NIHR_HIC_ICU_0043")
dt <- resolve_death()
}
