% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retrieve}
\alias{retrieve}
\title{Prepares workspace by retrieving all the POSTgres tables without collection}
\usage{
retrieve(connection)
}
\arguments{
\item{connection}{a formal PostgreSQL connection}
}
\description{
These tables come as a list and must be unlisted or accessed directly
}
