% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_parameters.R
\name{retrieve_unique_cases}
\alias{retrieve_unique_cases}
\title{Retrieve all unique cases}
\usage{
retrieve_unique_cases(episodes = NULL, provenance = NULL)
}
\arguments{
\item{episodes}{episode table}

\item{provenance}{provenance table}
}
\value{
a tibble with fields for site, nhs_number, episode_id and start
}
\description{
Pulls unique record numbers from the CC-HIC database. This requires only that an "episode" has
been correctly parsed from the xml and added to the database and so doesn't rely on, or infer that
there is any, event level data. This is the most accurate measure of the actual number of "patients"
in the system.
}
\examples{
retrieve_unique_cases(episodes, provenance)
}
