% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_int.R
\name{summarise_int}
\alias{summarise_int}
\title{Provides a summary for time-series data of "real"/"double" data type}
\usage{
summarise_int(processed_table = NULL, los_table = NULL, inUnits = "hours")
}
\arguments{
\item{los_table}{}

\item{inUnits}{character for duration based units to retunr}

\item{real_input}{character vector of NIHR codes for the reals that you wish to report}

\item{outliers}{integer value for the number of standard deviations used for outlier identification}
}
\value{
A tibble of time series data in summary
}
\description{
Provides a summary for time-series data of "real"/"double" data type
}
\examples{
\dontrun{summarise_dbl(c("events", "episodes", "provenance"), los, "NIHR_HIC_ICU_0108")}
}
