% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_string.R
\name{summarise_string}
\alias{summarise_string}
\title{Provides a summary for "string" data type}
\usage{
summarise_string(table_list = NULL, string_input = c("NIHR_HIC_ICU_0058",
  "NIHR_HIC_ICU_0098"))
}
\arguments{
\item{table_list}{a character vector for the uncollected tables of events, episodes and provenance (in that order)}

\item{string_input}{a character vector of the strings you want to extract}
}
\description{
Provides a summary for "string" data type
}
