% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_episode}
\alias{validate_episode}
\title{Validate CC-HIC Episodes}
\usage{
validate_episode(episode_length_tbl, invalidated_months)
}
\arguments{
\item{episode_length_tbl}{tibble containing episode characteristics from epi_length()}

\item{invalidated_months}{tibble containing months with sparse data contribution from invalid_months()}
}
\value{
a tibble with a single column containing episodes that have passed basic (level 1) validation checks
}
\description{
This function removes episodes that occur during particularly sparse periods (as this is likely that these
months are contributing poor data) and return only episodes that have a logical consistency in non-sparse months
}
