% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_validation.R
\name{validate_field}
\alias{validate_field}
\title{Validate HIC Field}
\usage{
validate_field(core_table = NULL, reference_table = NULL,
  input_field = NULL, qref = NULL, los_table = NULL)
}
\arguments{
\item{core_table}{core table}

\item{reference_table}{reference table}

\item{input_field}{character vecotr of input field}

\item{qref}{quality reference table}

\item{los_table}{episode length table}
}
\value{
a list with:
\enumerate{
  \item flagged events
  \item warning summaries
  \item
}
}
\description{
Wraps the following workflow:
Extract > Flag > Summarise > Plot
}
\examples{
validate_field(core, reference, "NIHR_HIC_ICU_0108", qref, episode_length)
}
