% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_parameters.R
\name{epi_length}
\alias{epi_length}
\title{Calculate Episode Length}
\usage{
epi_length(core = NULL, useDeath = TRUE)
}
\arguments{
\item{core}{core table from \code{make_core()}}
}
\description{
Calculates length of stay from:
\itemize{
  \item episode start date time: NHIC 0411
  \item episode end date timeL NHIC 0412
  \item date of in unit death: NHIC 0042
  \item time of in unit death: NHIC 0043
}
For cases where there is a discharge date, time and death date and time, we use the
mortality data if this occurs earlier than the discharge data.
}
\examples{
epi_length(core, reference)
}
