% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_demographics.R
\name{extract_demographics}
\alias{extract_demographics}
\title{Extract 1d data from CC-HIC}
\usage{
extract_demographics(metadata = NULL, events = NULL,
  codes = "NIHR_HIC_ICU_0093")
}
\arguments{
\item{metadata}{a database metadata table}

\item{events}{a database events table}

\item{codes}{a character vector of HIC codes you want to retrieve}
}
\value{
A tibble of 1d data
}
\description{
Extract 1d data from CC-HIC
}
\examples{
extract_demographics(tbls[["variables"]], tbls[["events"]])
}
