% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_2d.R
\name{extract_timevarying}
\alias{extract_timevarying}
\title{Reshapes timevarying events into hourly cadance table}
\usage{
extract_timevarying(events, metadata, chunk_size = 5000)
}
\arguments{
\item{events}{database events table}

\item{metadata}{database metadata table (collected)}
}
\value{
sparse tibble with hourly cadance as rows, and unique hic events as columns
}
\description{
Reshapes timevarying events into hourly cadance table
}
\examples{
\dontrun{
extract_timevarying(tbls[["events"]], tbls[["variables"]] \%>\% collect)
extract_timevarying(tbls[["events"]] \%>\%
filter(code_name == "NIHR_HIC_ICU_0108"), tbls[["variables"]] \%>\% collect) # returns just heart rates

}

}
