% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{ks_test}
\alias{ks_test}
\title{Pairwise Kolmogorov-Smirnov Test}
\usage{
ks_test(x)
}
\arguments{
\item{x}{extracted data item}

\item{site}{}
}
\description{
This takes an extract data item from CC-HIC and returns
the pairwise KS test by site.
}
\examples{
ks_test(heart_rates)
}
