% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_report.R
\name{make_report}
\alias{make_report}
\title{Make HIC Report Data}
\usage{
make_report(database = "lenient_dev", username = "edward",
  password = "superdb", path_name = NULL, useDeath = TRUE)
}
\arguments{
\item{path_name}{}
}
\description{
This is the main function for processing data from CC-HIC. It wraps all alther functions to evaluate
the whole database in one sweep.
}
\details{
When running this function you should make the following folder heirachy availible:
plots
working_data

report.rmd can be run after this function has successfully completed to produce a full report for CC-HIC
}
