% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendarHeat.R
\name{plot_heatcal}
\alias{plot_heatcal}
\title{Plot Calendar Heatmap}
\usage{
plot_heatcal(reference_table = NULL, site = NULL, filename = NULL,
  start_date = "2014-01-01", end_date = "2018-01-01")
}
\arguments{
\item{site}{character string of hospital site to plot}

\item{filename}{optional string to save output. must include extension}

\item{episodes}{collected episodes table}

\item{provenance}{collected provenance table}
}
\value{
a plot with admission details in heatmap form
}
\description{
Plot Calendar Heatmap
}
\examples{
plot_heatcal(episodes, provenance, "UCL")
plot_heatcal(episodes, provenance, "UCL", "~/some/path/plot.png")
}
