% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hic}
\alias{plot_hic}
\title{Plot NHIC Events}
\usage{
plot_hic(x, input = NULL)
}
\arguments{
\item{input}{the input variable of choice}

\item{core_table}{core table}

\item{valid_episodes}{an integer vector of episode IDs that have passed the validation process}

\item{qref}{the quality reference table}
}
\value{
A tibble 1 row per event
}
\description{
Plots NHIC events by a pre-determined plot
}
