% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_report.R
\name{report_cases_byday}
\alias{report_cases_byday}
\title{Reports Case Numbers by Site and Day}
\usage{
report_cases_byday(unique_cases_tbl = NULL, bysite = NULL)
}
\arguments{
\item{unique_cases_tbl}{from \code{pull_cases_all}}

\item{site}{}
}
\description{
Reports Case Numbers by Site and Day
}
