% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_parameters.R
\name{resolve_date_time}
\alias{resolve_date_time}
\title{Calculate Date and time of events}
\usage{
resolve_date_time(core_tbl = NULL, date_code = as.character(NULL),
  time_code = as.character(NULL))
}
\arguments{
\item{time_code}{}

\item{core}{core table from \code{make_core()}}
}
\value{
a function to encapsulate paired date and time variables stored separately
}
